cleaned utils and began redoing the cleaning of the data
